+++ openeo/metadata.py
@@ -213,6 +213,19 @@ class BandDimension(Dimension):
     def rename(self, name) -> Dimension:
         return BandDimension(name=name, bands=self.bands)
 
+
+class GeometryDimension(Dimension):
+    # TODO: how to model/store labels of geometry dimension?
+    def __init__(self, name: str):
+        super().__init__(name=name, type="geometry")
+
+    def rename(self, name) -> Dimension:
+        return GeometryDimension(name=name)
+
+    def rename_labels(self, target, source) -> Dimension:
+        return GeometryDimension(name=self.name)
+
+
 class CubeMetadata:
     """
     Interface for metadata of a data cube.
@@ -229,7 +242,7 @@ class CubeMetadata:
         if dimensions is not None:
             for dim in self._dimensions:
                 # TODO: here we blindly pick last bands or temporal dimension if multiple. Let user choose?
-                # TODO: add spacial dimension handling?
+                # TODO: add spatial dimension handling?
                 if dim.type == "bands":
                     if isinstance(dim, BandDimension):
                         self._band_dimension = dim
@@ -283,6 +296,16 @@ class CubeMetadata:
     @property
     def spatial_dimensions(self) -> List[SpatialDimension]:
         return [d for d in self._dimensions if isinstance(d, SpatialDimension)]
+
+    def has_geometry_dimension(self):
+        return any(isinstance(d, GeometryDimension) for d in self._dimensions)
+
+    @property
+    def geometry_dimension(self) -> GeometryDimension:
+        for d in self._dimensions:
+            if isinstance(d, GeometryDimension):
+                return d
+        raise MetadataException("No geometry dimension")
 
     @property
     def bands(self) -> List[Band]:
@@ -365,7 +388,7 @@ class CubeMetadata:
         dimensions = [d for d in self._dimensions if not isinstance(d, SpatialDimension)]
         return self._clone_and_update(dimensions=dimensions)
 
-    def add_dimension(self, name: str, label: Union[str, float], type: str = None) -> CubeMetadata:
+    def add_dimension(self, name: str, label: Union[str, float], type: Optional[str] = None) -> CubeMetadata:
         """Create new CubeMetadata object with added dimension"""
         if any(d.name == name for d in self._dimensions):
             raise DimensionAlreadyExistsException(f"Dimension with name {name!r} already exists")
@@ -375,6 +398,8 @@ class CubeMetadata:
             dim = SpatialDimension(name=name, extent=[label, label])
         elif type == "temporal":
             dim = TemporalDimension(name=name, extent=[label, label])
+        elif type == "geometry":
+            dim = GeometryDimension(name=name)
         else:
             dim = Dimension(type=type or "other", name=name)
         return self._clone_and_update(dimensions=self._dimensions + [dim])